requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
decorator==4.0.10         # via pycontracts
dicttoxml==1.7.4
django-extensions==1.7.4
django-memoize==1.2.0
django-mptt==0.8.6
django-reversion-compare==0.7.1
django-reversion==2.0.6
Django==1.9.9             # via django-memoize, django-reversion, django-reversion-compare
psycopg2==2.6.1
PyContracts==1.7.9
pyparsing==2.1.9          # via pycontracts
schematics==1.1.1
six==1.10.0               # via django-extensions, pycontracts, schematics
